+++ src/components/LinksPanel/components/index.ts
@@ -3,5 +3,6 @@ export * from './ContentElement';
 export * from './LinkElement';
 export * from './LinksGridLayout';
 export * from './LinksLayout';
+export * from './LoadingBar';
 export * from './MenuElement';
 export * from './TimePickerElement';